+++ build.gradle
@@ -39,50 +39,40 @@ subprojects {
     }
 }
 
-project(':kitchenpos-common') {
-    dependencies {
-    }
-}
-
 project(':kitchenpos-product') {
     dependencies {
-        compile project(':kitchenpos-common')
-        testCompile project(':kitchenpos-application').sourceSets.test.output
+        implementation project(':kitchenpos-common')
     }
 }
 
 project(':kitchenpos-menu') {
     dependencies {
-        compile project(':kitchenpos-common')
-        compile project(':kitchenpos-product')
-        testCompile project(':kitchenpos-application').sourceSets.test.output
+        implementation project(':kitchenpos-common')
+        implementation project(':kitchenpos-product')
     }
 }
 
 project(':kitchenpos-order') {
     dependencies {
-        compile project(':kitchenpos-common')
-        compile project(':kitchenpos-menu')
-        testCompile project(':kitchenpos-application').sourceSets.test.output
-        testCompile project(':kitchenpos-table')
+        implementation project(':kitchenpos-common')
+        implementation project(':kitchenpos-menu')
     }
 }
 
 project(':kitchenpos-table') {
     dependencies {
-        compile project(':kitchenpos-common')
-        compile project(':kitchenpos-order')
-        compile project(':kitchenpos-menu')
-        testCompile project(':kitchenpos-application').sourceSets.test.output
+        implementation project(':kitchenpos-common')
+        implementation project(':kitchenpos-order')
+        implementation project(':kitchenpos-menu')
     }
 }
 
 project(':kitchenpos-application') {
     dependencies {
-        compile project(':kitchenpos-common')
-        compile project(':kitchenpos-order')
-        compile project(':kitchenpos-menu')
-        compile project(':kitchenpos-product')
-        compile project(':kitchenpos-table')
+        implementation project(':kitchenpos-common')
+        implementation project(':kitchenpos-order')
+        implementation project(':kitchenpos-menu')
+        implementation project(':kitchenpos-product')
+        implementation project(':kitchenpos-table')
     }
 }
